split encoding into v2, cl, and helpers submodules
