+++ packages/sdk/src/swap/providers/velodrome/encoding/v2.ts
@@ -1,16 +1,8 @@
 import type { Address, Hex, PublicClient } from 'viem'
-import {
-  encodeAbiParameters,
-  encodeFunctionData,
-  encodePacked,
-  formatUnits,
-} from 'viem'
+import { encodeAbiParameters, encodeFunctionData, encodePacked } from 'viem'
 
-import { WETH } from '@/constants/assets.js'
 import type { SupportedChainId } from '@/constants/supportedChains.js'
 import {
-  CL_POOL_FACTORY_ABI,
-  CL_QUOTER_ABI,
   LEAF_ROUTER_ABI,
   POOL_ABI,
   POOL_FACTORY_ABI,
@@ -20,70 +12,12 @@ import {
 import type { VelodromeRouterType } from '@/swap/providers/velodrome/addresses.js'
 import type { Asset } from '@/types/asset.js'
 import type { SwapPrice, SwapRoute } from '@/types/swap/index.js'
-import { getAssetAddress, isNativeAsset } from '@/utils/assets.js'
+import { isNativeAsset } from '@/utils/assets.js'
+
+import { buildSwapPrice, MSG_SENDER, resolveTokens } from './helpers.js'
 
 // ─────────────────────────────────────────────────────────────────────────────
-// Shared helpers
-// ─────────────────────────────────────────────────────────────────────────────
-
-/** Sentinel: route output to msg.sender */
-const MSG_SENDER = '0x0000000000000000000000000000000000000001' as Address
-
-/** Resolve asset pair to on-chain token addresses for a given chain. */
-function resolveTokens(
-  assetIn: Asset,
-  assetOut: Asset,
-  chainId: SupportedChainId,
-): { tokenIn: Address; tokenOut: Address } {
-  const tokenIn = isNativeAsset(assetIn)
-    ? getWrappedNativeAddress(chainId)
-    : getAssetAddress(assetIn, chainId)
-  const tokenOut = isNativeAsset(assetOut)
-    ? getWrappedNativeAddress(chainId)
-    : getAssetAddress(assetOut, chainId)
-  return { tokenIn, tokenOut }
-}
-
-/**
- * Get the wrapped native token address for a chain.
- * Velodrome routers require WETH address in Route structs, not address(0).
- */
-function getWrappedNativeAddress(chainId: SupportedChainId): Address {
-  const addr = WETH.address[chainId]
-  if (!addr || addr === 'native') {
-    throw new Error(`No WETH address configured for chain ${chainId}`)
-  }
-  return addr
-}
-
-/** Build a SwapPrice from raw quote data. */
-function buildSwapPrice(
-  assetIn: Asset,
-  assetOut: Asset,
-  amountInRaw: bigint,
-  amountOutRaw: bigint,
-  route: SwapRoute,
-): SwapPrice {
-  const amountIn = parseFloat(
-    formatUnits(amountInRaw, assetIn.metadata.decimals),
-  )
-  const amountOut = parseFloat(
-    formatUnits(amountOutRaw, assetOut.metadata.decimals),
-  )
-  return {
-    price: (amountOut / amountIn).toFixed(6),
-    priceInverse: (amountIn / amountOut).toFixed(6),
-    amountIn,
-    amountOut,
-    amountInRaw,
-    amountOutRaw,
-    priceImpact: 0,
-    route,
-  }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// V2 AMM quoting
+// Quoting
 // ─────────────────────────────────────────────────────────────────────────────
 
 export interface GetQuoteParams {
@@ -114,7 +48,14 @@ export async function getQuote(params: GetQuoteParams): Promise<SwapPrice> {
   return buildSwapPrice(assetIn, assetOut, amountInRaw, amountOutRaw, route)
 }
 
-/** Fetch the output amount using the appropriate quoting mechanism for the router type. */
+/**
+ * Fetch the output amount using the appropriate quoting mechanism for the router type.
+ * @param params - Quote parameters
+ * @param tokenIn - Resolved input token address
+ * @param tokenOut - Resolved output token address
+ * @returns Output amount as raw bigint
+ * @throws If router type is unknown
+ */
 async function fetchAmountOut(
   params: GetQuoteParams,
   tokenIn: Address,
@@ -143,7 +84,7 @@ async function fetchAmountOut(
   throw new Error(`Unknown router type: ${routerType as string}`)
 }
 
-/** Quote via Pool.getAmountOut (Universal Router path — no legacy router available). */
+/** Quote via Pool.getAmountOut (Universal Router path — no router-level quoting available). */
 async function fetchAmountOutViaPool(
   params: GetQuoteParams,
   tokenIn: Address,
@@ -178,8 +119,8 @@ async function fetchAmountOutViaPool(
 /** Quote via Router.getAmountsOut (v2 and leaf router path). */
 async function fetchAmountOutViaRouter(
   params: GetQuoteParams,
-  tokenIn: Address,
-  tokenOut: Address,
+  _tokenIn: Address,
+  _tokenOut: Address,
   abi: typeof V2_ROUTER_ABI | typeof LEAF_ROUTER_ABI,
   route: { from: Address; to: Address; stable: boolean; factory?: Address },
 ): Promise<bigint> {
@@ -193,7 +134,7 @@ async function fetchAmountOutViaRouter(
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
-// V2 AMM encoding
+// Encoding
 // ─────────────────────────────────────────────────────────────────────────────
 
 export interface EncodeSwapParams {
@@ -213,6 +154,7 @@ export interface EncodeSwapParams {
  * Encode swap calldata for the appropriate router type.
  * @param params - Swap encoding parameters
  * @returns Encoded calldata as hex string
+ * @throws If router type is unknown
  */
 export function encodeSwap(params: EncodeSwapParams): Hex {
   const { routerType } = params
@@ -226,7 +168,7 @@ export function encodeSwap(params: EncodeSwapParams): Hex {
     return encodeUniversalV2Swap(tokenIn, tokenOut, params)
   }
   if (routerType === 'v2') {
-    return encodeLegacySwap(tokenIn, tokenOut, params, V2_ROUTER_ABI, {
+    return encodeRouterSwap(tokenIn, tokenOut, params, V2_ROUTER_ABI, {
       from: tokenIn,
       to: tokenOut,
       stable: params.stable,
@@ -234,7 +176,7 @@ export function encodeSwap(params: EncodeSwapParams): Hex {
     })
   }
   if (routerType === 'leaf') {
-    return encodeLegacySwap(tokenIn, tokenOut, params, LEAF_ROUTER_ABI, {
+    return encodeRouterSwap(tokenIn, tokenOut, params, LEAF_ROUTER_ABI, {
       from: tokenIn,
       to: tokenOut,
       stable: params.stable,
@@ -287,10 +229,10 @@ function encodeUniversalV2Swap(
 }
 
 /**
- * Encode swap calldata for legacy v2 or leaf routers.
+ * Encode swap calldata for v2 or leaf routers.
  * Selects the correct swap function based on whether native ETH is involved.
  */
-function encodeLegacySwap(
+function encodeRouterSwap(
   tokenIn: Address,
   tokenOut: Address,
   params: EncodeSwapParams,
@@ -320,145 +262,5 @@ function encodeLegacySwap(
     abi,
     functionName: 'swapExactTokensForTokens',
     args: [amountInRaw, amountOutMin, [route], recipient, BigInt(deadline)],
-  })
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// CL / Slipstream quoting
-// ─────────────────────────────────────────────────────────────────────────────
-
-export interface GetCLQuoteParams {
-  assetIn: Asset
-  assetOut: Asset
-  amountInRaw: bigint
-  chainId: SupportedChainId
-  publicClient: PublicClient
-  clFactoryAddress: Address
-  clQuoterAddress: Address
-  tickSpacing: number
-}
-
-/**
- * Get a swap quote from a CL/Slipstream pool via QuoterV2.
- * Verifies the pool exists via the CL factory, then quotes via QuoterV2.
- * @param params - CL quote parameters
- * @returns Price quote with amounts and route
- * @throws If no CL pool exists for the given pair and tickSpacing
- */
-export async function getCLQuote(params: GetCLQuoteParams): Promise<SwapPrice> {
-  const {
-    assetIn,
-    assetOut,
-    amountInRaw,
-    publicClient,
-    clFactoryAddress,
-    clQuoterAddress,
-    tickSpacing,
-    chainId,
-  } = params
-  const { tokenIn, tokenOut } = resolveTokens(assetIn, assetOut, chainId)
-
-  // Verify the CL pool exists
-  const poolAddress = await publicClient.readContract({
-    address: clFactoryAddress,
-    abi: CL_POOL_FACTORY_ABI,
-    functionName: 'getPool',
-    args: [tokenIn, tokenOut, tickSpacing],
-  })
-
-  if (
-    !poolAddress ||
-    poolAddress === '0x0000000000000000000000000000000000000000'
-  ) {
-    throw new Error(
-      `No CL pool found for ${assetIn.metadata.symbol}/${assetOut.metadata.symbol} (tickSpacing=${tickSpacing})`,
-    )
-  }
-
-  // Quote via QuoterV2.quoteExactInputSingle
-  // sqrtPriceLimitX96 = 0 means no price limit
-  const quoteResult = (await publicClient.readContract({
-    address: clQuoterAddress,
-    abi: CL_QUOTER_ABI,
-    functionName: 'quoteExactInputSingle',
-    args: [
-      {
-        tokenIn,
-        tokenOut,
-        amountIn: amountInRaw,
-        tickSpacing,
-        sqrtPriceLimitX96: 0n,
-      },
-    ],
-  })) as readonly [bigint, bigint, number, bigint]
-
-  const route: SwapRoute = {
-    path: [assetIn, assetOut],
-    pools: [{ address: poolAddress as Address, fee: 0, version: 'v3' }],
-  }
-  return buildSwapPrice(assetIn, assetOut, amountInRaw, quoteResult[0], route)
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// CL / Slipstream encoding
-// ─────────────────────────────────────────────────────────────────────────────
-
-export interface EncodeCLSwapParams {
-  assetIn: Asset
-  assetOut: Asset
-  amountInRaw: bigint
-  amountOutMin: bigint
-  tickSpacing: number
-  recipient: Address
-  deadline: number
-  chainId: SupportedChainId
-}
-
-/** Universal Router V3_SWAP_EXACT_IN command byte */
-const V3_SWAP_EXACT_IN = 0x00
-
-/**
- * Encode a V3_SWAP_EXACT_IN command for a CL/Slipstream pool on the Universal Router.
- * Path: encodePacked([tokenIn (20), tickSpacing as int24 (3), tokenOut (20)]) — 43 bytes.
- * @param params - CL swap encoding parameters
- * @returns Encoded calldata as hex string
- */
-export function encodeCLSwap(params: EncodeCLSwapParams): Hex {
-  const { amountInRaw, amountOutMin, tickSpacing, deadline, chainId } = params
-  const { tokenIn, tokenOut } = resolveTokens(
-    params.assetIn,
-    params.assetOut,
-    chainId,
-  )
-
-  const commands = `0x${V3_SWAP_EXACT_IN.toString(16).padStart(2, '0')}` as Hex
-
-  // CL path: [tokenIn (20)] [tickSpacing as int24 (3)] [tokenOut (20)] — 43 bytes
-  const path = encodePacked(
-    ['address', 'int24', 'address'],
-    [tokenIn, tickSpacing, tokenOut],
-  )
-
-  const input = encodeAbiParameters(
-    [
-      { type: 'address' },
-      { type: 'uint256' },
-      { type: 'uint256' },
-      { type: 'bytes' },
-      { type: 'bool' },
-    ],
-    [
-      MSG_SENDER,
-      amountInRaw,
-      amountOutMin,
-      path,
-      false, // payerIsUser — tokens pre-transferred to router
-    ],
-  )
-
-  return encodeFunctionData({
-    abi: UNIVERSAL_ROUTER_ABI,
-    functionName: 'execute',
-    args: [commands, [input], BigInt(deadline)],
   })
 }
